Added my test folder from hashtag pull, fixed file copy code.
Currently the program will copy files over to a new directory if they have only one face in them. Next, I'm going to fix this JSON import, I ran into a snag that I need to debug but I think it might be the JSON file and not the code. On to the next challenge :)