This file allows you to import a json file, identify the image name, and search each image for faces.
I'm going to give this a once over, but basically the source directory suffix is where your files will come from, the destination is where all files that have one face are copied to, and the json file is the location of the json file.

My TODO other than cleaning is to make sure to include/exclude based on hashtags!